OMAP: PM: Allowing an early init of pm debugfs driver.
This patch changes the pm_db_init from arch initcall to a postcore
initcall. With arch initcall, it is impossible for pm driver that
gets initialized prior to this driver to use one of the
pm debug fs entries during its init. Making it a postcore initcall
ensures that this drver gets initialized early on before any pm
drivers.

Signed-off-by: Thara Gopinath <thara@ti.com>