OMAP2: OPP: update API to be device-based
The current OPP layer is based on a unique identifier (OPP_MPU,
OPP_DSP, OPP_L3) which is not terribly portable or scalable for future
devices.

Since we'd also like to be able to manage OPPs for any device, this
patch changes the OPP API to be device-based instead of unique-ID
based.  Essentially, this means passing a 'struct device' instead of a
unique ID to determine which set of OPPs to be used.

While making the API change, I also significantly changed the
internal management of OPPs to use a list per-device and within
each per-device struct, a list to manage OPPs.  This resulted in
basically a rewrite of the internals.

Another signifcant change is that I removed the opp_init_list()
function in favor of just using multiple calls to opp_add().  In
combination with the internal list-based implmentation, this permitted
the removal of the need for using termin

Cc: Nishanth Menon <nm@ti.com>
Signed-off-by: Kevin Hilman <khilman@deeprootsystems.com>