TILER: Added support for granularity and in-tile offset.
Offset was cropped to tile boundaries.  Now supporting full offset.

Granularity and alignment was combined as the same concept.  Now
we support (min.) granularity even for larger aligmnents.

Granularity is the minimum TILER block size, which could be aligned
to a larger alignment value.  This prevents allocating too much
TILER space in this case.

Signed-off-by: Lajos Molnar <molnar@ti.com>
Signed-off-by: Sriramakrishnan A G <srk@ti.com>