OMAP: Remove dependency of generic opp layer on cpufreq.
This patch removes the dependency of the opp layer on cpufreq layer.
OPP layer is now enabled to compile and exist in the system
irrespective of whether cpu freq layer is enabled or not.

[NM Note:
Original Patch: https://patchwork.kernel.org/patch/111664/
Still pending: http://marc.info/?l=linux-omap&m=127530955314582&w=2
To be squashed to patch 1
]

Signed-off-by: Thara Gopinath <thara@ti.com>