OMAP3: PM: Adding debug support to Voltage and Smartreflex drivers
This patch adds debug support to the voltage and smartreflex drivers.
This means a whole bunch of voltage processor and smartreflex
parameters are now visible through the pm debugfs. By default
only a read of these parameters are permitted. If you need to
write into them then
	echo 1 > /pm_debug/enable_sr_vp_debug

The voltage parameters can be viewed at
	/pm_debug/voltage/vdd_<x>/<parameter>
and the smartreflex parameters can be viewed at
	/pm_debug/smartreflex/sr_<x>/<parameter>

where <x> is mpu or core for OMAP3.

Signed-off-by: Thara Gopinath <thara@ti.com>