OMAP3: disable idle during the entire suspend/resume sequence
Idle path should be disabled during the entire suspend/resume sequence.
Currently it is disabled in ->prepare() and re-enabled in ->finish(),
but the suspend sequence starts with ->begin() and ends with ->end(),
leaving windows where the suspend/resume sequence is still underway and
idle code could execute.

The fix is to move the idle disable (disable_hlt) and enable (enable_hlt)
into ->begin() and ->end() respectively to ensure that the idle path
is disabled for the entire suspend/resume sequence.
Removing omap3_pm_prepare and omap3_pm_finish since they are now empty.

Tested with RET and OFF on Beagle and OMAP3EVM.

Changelog rework as suggested by Kevin.

Signed-off-by: Jean Pihet <j-pihet@ti.com>
Signed-off-by: Kevin Hilman <khilman@deeprootsystems.com>