TI814X: clock: Add clockdomains and powerdomains data
This patch adds data for various clock domains and power domains in TI814X.

The data common between TI816X and TI814X is updated with TI81XX/ti81xx prefix
with dependent code modified as applicable.

Note that at present this is not exhaustive and need to add missing domains.

Signed-off-by: Anil Kumar Ch <anilkumar@ti.com>
Signed-off-by: Hemant Pedanekar <hemantp@ti.com>