omap3: beaglexm: power on DVI
TFP410 DVI chip is used to provide display out.
this chip is controlled by 2 lines:
LDO which supplies the power is controlled over gpio + 2
and the enable of the chip itself is done over gpio + 1
NOTE: the LDO is necessary for LED, serial blocks as well.

GPIO + 1 was used to sense USB overcurrent in vanilla beagle

[nm@ti.com: helped split up the patch]
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Koen Kooi <koen@beagleboard.org>
Signed-off-by: Kevin Hilman <khilman@ti.com>