ti816x: Added Smart Reflex Class 2 driver
This patch adds the Smart Reflex driver to TI816X processor.
Adaptive Voltage Scaling is done only for core in TI816X.

Voltage is changed based on both SVT and HVT sensor values.
Voltage is changed at IRQ handling

Added necessary clocks to the clock tree of TI816X for enabling
the adddress space of SR module and also modified the KConfig
which gives us an option to include or not-include Smart Reflex
driver to the Kernel.

Signed-off-by: Anil Kumar Ch <anilkumar@ti.com>