ARM: 6674/1: LPAE: use long long format when printing physical addresses and ptes
For the Kernel to support 2 level and 3 level page tables, physical
addresses (and also page table entries) need to be 32 or 64-bits depending
upon the configuration.

This patch uses the %08llx conversion specifier for physical addresses
and page table entries, ensuring that they are cast to (long long) so
that common code can be used regardless of the datatype widths.

Acked-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>

This patch was imported from linux-omap git commit id
29a38193c15bd72ba96c57a805443ef46ea6a6ac. Since it was on latest kernel (3.x),
removed the files which were not needed to be patched in 2.6.37 to avoid
warnings.

In addition, removed similar warning form plat-omap/sram.c due to usage of
phys_addr_t in __pfn_to_phys().

Signed-off-by: Hemant Pedanekar <hemantp@ti.com>