vpfe_capture: Remove INIT_LIST_HEAD from .buf_release
This seems to be bug in driver, this could have been introduced due
to some of major changes in videobuf layes.

We must not do INIT_LIST_HEAD in buf_release function, since for every
qbuf the buf_release function gets called if address is not same for the
given perticular index. Due to this, the the queue will get reset to init
state irrespective of buffer state.

Signed-off-by: Sriramakrishnan A G <srk@ti.com>