TI816X: gpio voltage regulator
This patch adds the gpio controlled voltage regulator, Which is used
to control the AVS voltage domain on TI816X.

Added some platform specific data to the board init file, which is independent
from board to board. Data includes voltage table, no. of voltages, regulator
init data, gpios & no. of gpio pins and etc

Voltage will change based on the gpio value provided to the driver from
others drivers, which are requested for a voltage change.

Signed-off-by: Anil Kumar Ch <anilkumar@ti.com>
Signed-off-by: Sriramakrishnan A G <srk@ti.com>