SYSLINK: multiproc Added support for omap3
multiproc module provides the ability to uniquely specify and
identify processors in a multi-processor environment.
With this patch, each processor in omap3 can  be uniquely
identified by either a name string or an integer ranging from 0 to 1.
Notify and SysLink modules require it to identify
remote cores and local core.

Signed-off-by: Ravindranath Andela <ravindranath.a@ti.com>
Signed-off-by: Sriramakrishnan A G <srk@ti.com>