TI81XX: MMC: This patch enables card detection for TI81XX
This patch enables the interrupt status enable registers for card
removal and insertion. Once an interrupt is obtained, this patch
adds support for handling it in case of card removal. Once the card
was removed, the SD bus power used to go off and therefore subsequent
card insertions could not be detected completely in software. This
patch handles this by enabling the SD bus power so that subsequent
card insetions can happen properly in software.

Signed-off-by: Saxena, Parth <parth.saxena@ti.com>
Signed-off-by: Sriramakrishnan A G <srk@ti.com>