ti816x: gpio: Added api for set array output
This patch adds one more API to GPIO lib for setting array in a
single write. gpio_direction_output_array() is the API, can be
used for setting array output. This is supported only when all
the GPIO pins are belonging to a single bank.

Also modified Voltage regulator code for using the above API for
writing GPIO value at a time to the register.

Signed-off-by: Anil Kumar Ch <anilkumar@ti.com>