ti814x: mux: Add pinmux support
This patch adds pimux data for TI814x and plugs into OMAP pinmux framework.

Following major differences with respect to OMAP are taken care of:
1) Mux modes are bit positions instead of 3 bit value
2) Pad config registers exceed 16 bit value and thus 32-bit values are used.
3) It is mandatory to do a read-modify-write on padconfig registers as some of
the reserved bits are writable but need to be maintained at power-on-reset
values (unless the user really wants to touch them).

TODO:
1) Add support for showing correct value of Pull Up/Down status through debug fs
for TI81xx.
2) Rename existing file/function/data for TI816x with TI816X/ti816x prefix
instead of TI81XX/ti81xx since the mux data and configuration is not common
across TI816x and TI814x.

Signed-off-by: Hemant Pedanekar <hemantp@ti.com>
Signed-off-by: Sriramakrishnan A G <srk@ti.com>