ti814x: mxt224 driver changes for CHG line acknowledge.
This patch moves the dummy read function after the reset.
The CHG line has to be pulled high before doing any operation.
This is done by doing dummy read of the message object.
Previously, dummy read was done before the reset. The issue
with that was, that even though the CHG line was driven HIGH
by dummy read function, the subsequent reset command made the
CHG line to go LOW again. This patch solves this by doing the
dummy read(CHG line High) after the reset.

Signed-off-by: Sugumar Natarajan <sugumar@ti.com>