ARM: add an extra temp register to the low level debugging addruart macro
Some platforms (like OMAP not to name it) are doing rather complicated
hacks just to determine the base UART address to use.  Let's give their
addruart macro some slack by providing an extra work register which will
allow for much needed cleanups.

This is basically a no-op as this commit is only adding the extra argument
to the macro but no one is using it yet.

Signed-off-by: nicolas Pitre <nicolas.pitre@linaro.org>
Reviewed-by: Kevin Hilman <khilman@ti.com>

This patch was imported from linux-omap git and was originally for 3.x kernel.
Removed conflicts and files not related to ARM (general) and OMAP.

The original commit id is 88af40e472c6f92469751567f88357a43fb8403e.

Signed-off-by: Hemant Pedanekar <hemantp@ti.com>