DM81xx V4L2 Capture: IR#SDOCM00085999 Solved.
HeadLine: Capture driver continously reports
"Unable to allocatebuffers"

Detailed Description:
This problem occurs only when trying to configure capture
driver with non-supported configuration. This is not
happening under normal conditions. Once capture driver
is configured for non-supported configuration it says
"unable to allocate buffers" After that user has to
restart the board and try to configure capture with proper values.