regulator: tps65910: Add voltage steps and fix logic for VDD1/2, VDDCtrl for TPS65911
The VDD1/2 and VDDCtrl on TPS65911 allow steps in 12.5 mV. Calculate the final
voltage run time but put a static entry for number of voltages possible for the
same. This allows qyerying numbers of voltages without which the voltage
selection will not be invoked form the core.

Also fix the check for error return when getting the voltage depending uping the
selector.

Signed-off-by: Hemant Pedanekar <hemantp@ti.com>