ti816x: clock: set rate function should not check for zero use count
Fix ti816x_clksel_set_rate function not to check for usecount zero
which would force a clock to be disabled whenever the rate
rate needs to be changed.
Instead, check for usecount to be '1' which would allow the user
to change the rate as and when needed. User has to get a pointer
to clock and enbales it once. Enabling once is mandated to identify
conflicts such as two users trying to change rate simultaneously.

NOTE: Earlier to this patch ti816x_clk_set_rate() function would dynamically
select a parent with usecount zero when current parent has usecount
greater than zero and change the rate. This feature has been removed
it might cause conflicts and rate changes without notification.

This change might cause some problem to drivers if they are sharing
the parent with other drivers and try to change the clock rate.
If any driver needs to change the rate then it should make sure no
other driver is using the clock/parent clock.

Signed-off-by: Ravikumar Kattekola <rk@ti.com>