ti814x: PM: Do not initialize SRAM on PG 1.x
Suspend resume code is pushed to SRAM(ARM internal RAM)
during init and used during suspend so that ddr can be put
in self-refresh.
As suspend to memory is not supported on PG 1.x we do not
need to initialize SRAM.
If SRAM is required then use OCMC RAM on PG 1.x
because ARM internal RAM is secured on PG 1.x and causes aborts
when accessed.

Signed-off-by: Ravikumar Kattekola <rk@ti.com>