ti814x: gpio: fix debounce clock issue
This patch fixes the debounce clock issue. Issue details are
mentioned below

Non-Zero value of gpio debounce_interval creates kernel hang.
This is due to debounce clock is not added to the gpio dbck clock
data. So while enabling the debounce clock it is not setting the
enable bit of GPIO debounce register. Due to this GPIO write on
to the OMAP4_GPIO_DEBOUNCINGTIME register is failing

Signed-off-by: AnilKumar Ch <anilkumar@ti.com>