IR#SDOCM00081265: Enable multiple VENC Simultaneously causing problem
When enable multiple venc simultaneoulsy, display showes flickers since
This will require the VPDMA to process the descriptors for all the vencs
almost at same time. Since VPDMA can process limited descriptors any of
the Venc will start under flowing and this will result in flicker.
What we did is that in fbdev we are enabling the Venc with some delay
this will allow the VPDMA to process descriptors at different times. Same
thing is done in OMX code.

This came out to be a bug in software in M3 side code. IR for same is
SDOCM00088589