usb: musb: cppi41: fix video capture issue
Fixes video capture issue with Logitech-prp9000 for 640X460 resolution.
The issues was with cppi41 dma programming for transparent mode with
high bandwidth endpoints where it was not taking care of multiple
maxpacket size data within one micro frame.

Added hb_mult to cppi41_channel structure to take care of such scenarios.

This patch is still being tested and not intended for any official release to
customer. This will be released in PSP release after complete testing.

Signed-off-by: Ajay Kumar Gupta <ajay.gupta@ti.com>
Signed-off-by: Visuvanadan Pasupathy <a0393168@psplinux064.(none)>