TI81xx:Lirc support for TI81xx
Modify lirc_serial driver to include CIR
support in TI81xx. Support added for NEC, SONY, RC-5.
Platform specific patch for pcf8575 IO Expander
need to be applied along with this to have full support for
CIR mode on required platform.
The reason for modification of this lirc_serial is that previously
the DCD pin of the UART was used to detect the signal change,
but in case of TI hardware, the RxD pin of UART is connected to
the IR receiver instead of DCD pin. So to detect the signal change,
these changes need to be done.

Signed-off-by: Harman Sohanpal <harman_sohanpal@ti.com>