TI81xx Capture: IR: Kernel dump is seen on starting capture
IR#SDOCM00085993
Headline: Kernel dump is seen on starting capture
Description: Kernel dump is seen on start of capture. This is because
VPSS driver is having a udelay of 50-100us. and with this delay we
take spinlock. so IRQ remains disable for long period of time.
So kernel is having dump to warn us. This requires a design
change to remove udelay and move to interrupt state.

Resolution: We have moved the code such that spinlock is not required
when udelay is required.