TI81xx V4L2 capture: Change filter mode based on preset_detected
IR#SDOCM00086261: Solved

HeadLine: Filter is always selected for 1080P60 mode
Description:
There is a filter between TVP7002 and VIP capture port.
Filter needs to be set based on selection of the input capture mode.
Currently filter is always selected for 1080P60 mode.
No artifacts are seen on any of the boards used for testing.
But ideally filter should be selected based on mode.