usb: musb: cppi41: correct data toggle mismatch to fix extra IN token issue
Fixes extra IN token issue seen when a file is transferred from obe pendrive
to other. The same issue is also seen when a file is read from a usb pendrive
and same time an usb wifi module is connected and its ip interface is up.

Issue root caused to be with xDMA corrupting the data toggle bit when multiple
bus masters are accessing the mentor core register space.

Signed-off-by: Ajay Kumar Gupta <ajay.gupta@ti.com>