TI81xx capture: Add i2c_adapter_id to subdev info
This patch adds i2c_adapter_id as a part of subdev info,
so that each subdevice can be connected on a different
i2c bus.
This enhancement is done as a part of IR#SDOCM00087388.
HeadLine: [HDVPSS-Capture] Adapter on which slave is
connected should be a platform device field.
Description: Various boads may have slave devices connected
to different I2C adapter preset on SoCs. So adapaters on
which these slaves are connected should be a platform
device field. Currently code assumes that slave is on I2C2
for TI816x and on I2C3 for TI814. This should ideally come
from platform data.

For the ease of connecting subdevices on different I2C
buses, i2c_adpater id is made subdev dependent.

Signed-off-by: Deepthy Ravi <deepthy.ravi@ti.com>