ti81xx: nand: fix 8-bit bch ecc scheme while reading a page
GPMC requires 13 bytes of ECC for every 512 bytes to implement
8-bit BCH scheme in the hardware. But ROM code uses 14 bytes of
ECC for every 512 bytes. So, to align our Uboot and Kernel with
the ROM code (so that the ECC layout is common throughout),
we implement 14 bytes of ECC for every 512 bytes of data and set
the 14th byte to 0 (since the GPMC requires only 13 bytes).

This patch accounts for considering 14 bytes of ECC for every
512 bytes of data (instead of 13 bytes), while reading a page.

Signed-off-by: Saxena, Parth <parth.saxena@ti.com>