ti811x: Add hwmod support for GPIO 4 and 5
TI811X has an additional two GPIO banks(4 and 5).
GPIO4/5 banks do not have dedicated irqs. TI811X has MPU/A8 Interrupt
multiplexing using Interrupt crossbar which provides MMRs(32-bit) to
configure the muxes. For details on event/int crossbar and muxes refer to
the 'Technical reference manual' for TI811X.

This patch choses irqs 95 to 92 for GPIO4/5 banks, by default these
are for events from TIMER4-7, please note that this is only for illustration.
The irqs to be used must be chosen based on the system configuration and
usecase so we do not override irqs of a module in use.

Signed-off-by: Ravikumar Kattekola <rk@ti.com>