drivers: net: cpsw_ale: Fix for VLAN delete entry
VLAN entry is not deleted as port mask comparison is invalid. ALE entry is
made free only when port mask equals to BIT(ale->ale_ports) which is 0x2
but an ALE entry must be freed only when port mask is 0x0.

This patch fixes the issue by making the ALE entry as free entry when port
mask is 0x0.

Signed-off-by: Mugunthan V N <mugunthanvnm@ti.com>