usb: musb: add CPPI41 DMA support
This commit adds support for CPPI41 kind of DMA channels for musb mode
of usb controllers. CPPI41 can be found in TI81XX processors family.

Files was taken from linux-omap3 kernel tree.

Also this DMA mode requires some changes to Linux's USB stack, that was
made by developers from TI. This change is not accurate, but we need
this and not pretend for in-kernel integration. For now.